refactor reexports
