--- src/feature/auth/ui/forgot-password-form/index.ts
@@ -1,1 +1,0 @@
-export { ForgotPasswordForm } from './forgot-password-form'
