+++ src/feature/auth/ui/index.ts
@@ -1,0 +1,5 @@
+export { CreatePasswordForm } from './create-password-form'
+export { ForgotPasswordForm } from './forgot-password-form'
+export { SentEmailModal } from './sent-email-modal'
+export { SignInForm } from './sign-in-form'
+export { SignUpForm } from './sign-up-form'
