--- src/feature/auth/ui/sent-email-modal/index.ts
@@ -1,1 +1,0 @@
-export { SentEmailModal } from './sent-email-modal'
