+++ src/feature/auth/ui/sent-email-modal/index.tsx
