--- src/feature/auth/ui/sign-in-form/index.ts
@@ -1,1 +1,0 @@
-export { SignInForm } from './sign-in-form'
