--- src/feature/auth/ui/sign-up-form/index.ts
@@ -1,1 +1,0 @@
-export { SignUpForm } from './sign-up-form'
